requirements.txt
# Production Requirements for Conversational Analytics System

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
asyncpg==0.29.0

# Data Processing
pandas==2.1.4
duckdb==0.9.2

# LLM & AI
groq==0.4.1
httpx==0.26.0

# SQL Parsing & Validation
sqlparse==0.4.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Production servers
gunicorn==21.2.0
